Split prompt context block builders
